+++ airbyte-cdk/python/airbyte_cdk/sources/declarative/models/declarative_component_schema.py
@@ -53,6 +53,23 @@ class CheckStream(BaseModel):
     )
 
 
+class ConcurrencyLevel(BaseModel):
+    type: Optional[Literal['ConcurrencyLevel']] = None
+    default_concurrency: Union[int, str] = Field(
+        ...,
+        description='The amount of concurrency that will applied during a sync. This value can be hardcoded or user-defined in the config if different users have varying volume thresholds in the target API.',
+        examples=[10, "{{ config['num_workers'] or 10 }}"],
+        title='Default Concurrency',
+    )
+    max_concurrency: Optional[int] = Field(
+        None,
+        description='The maximum level of concurrency that will be used during a sync. This becomes a required field when the default_concurrency derives from the config, because it serves as a safeguard against a user-defined threshold that is too high.',
+        examples=[20, 100],
+        title='Max Concurrency',
+    )
+    parameters: Optional[Dict[str, Any]] = Field(None, alias='$parameters')
+
+
 class ConstantBackoffStrategy(BaseModel):
     type: Literal['ConstantBackoffStrategy']
     backoff_time_in_seconds: Union[float, str] = Field(
@@ -666,6 +683,10 @@ class KeysToLower(BaseModel):
 
 class IterableDecoder(BaseModel):
     type: Literal['IterableDecoder']
+
+
+class XmlDecoder(BaseModel):
+    type: Literal['XmlDecoder']
 
 
 class MinMaxDatetime(BaseModel):
@@ -1304,6 +1325,7 @@ class DeclarativeSource(BaseModel):
     schemas: Optional[Schemas] = None
     definitions: Optional[Dict[str, Any]] = None
     spec: Optional[Spec] = None
+    concurrency_level: Optional[ConcurrencyLevel] = None
     metadata: Optional[Dict[str, Any]] = Field(
         None,
         description='For internal Airbyte use only - DO NOT modify manually. Used by consumers of declarative manifests for storing related metadata.',
@@ -1440,8 +1462,8 @@ class SessionTokenAuthenticator(BaseModel):
         description='Authentication method to use for requests sent to the API, specifying how to inject the session token.',
         title='Data Request Authentication',
     )
-    decoder: Optional[JsonDecoder] = Field(
-        None, description='Component decoding the response', title='Decoder'
+    decoder: Optional[Union[JsonDecoder, XmlDecoder]] = Field(
+        None, description='Component used to decode the response.', title='Decoder'
     )
     parameters: Optional[Dict[str, Any]] = Field(None, alias='$parameters')
 
@@ -1570,6 +1592,11 @@ class ParentStreamConfig(BaseModel):
         description='Indicates whether the parent stream should be read incrementally based on updates in the child stream.',
         title='Incremental Dependency',
     )
+    extra_fields: Optional[List[List[str]]] = Field(
+        None,
+        description='Array of field paths to include as additional fields in the stream slice. Each path is an array of strings representing keys to access fields in the respective parent record. Accessible via `stream_slice.extra_fields`. Missing fields are set to `None`.',
+        title='Extra Fields',
+    )
     parameters: Optional[Dict[str, Any]] = Field(None, alias='$parameters')
 
 
@@ -1607,10 +1634,12 @@ class SimpleRetriever(BaseModel):
         description='PartitionRouter component that describes how to partition the stream, enabling incremental syncs and checkpointing.',
         title='Partition Router',
     )
-    decoder: Optional[Union[JsonDecoder, JsonlDecoder, IterableDecoder]] = Field(
-        None,
-        description='Component decoding the response so records can be extracted.',
-        title='Decoder',
+    decoder: Optional[Union[JsonDecoder, JsonlDecoder, IterableDecoder, XmlDecoder]] = (
+        Field(
+            None,
+            description='Component decoding the response so records can be extracted.',
+            title='Decoder',
+        )
     )
     parameters: Optional[Dict[str, Any]] = Field(None, alias='$parameters')
 
@@ -1671,10 +1700,12 @@ class AsyncRetriever(BaseModel):
         description='PartitionRouter component that describes how to partition the stream, enabling incremental syncs and checkpointing.',
         title='Partition Router',
     )
-    decoder: Optional[Union[JsonDecoder, JsonlDecoder, IterableDecoder]] = Field(
-        None,
-        description='Component decoding the response so records can be extracted.',
-        title='Decoder',
+    decoder: Optional[Union[JsonDecoder, JsonlDecoder, IterableDecoder, XmlDecoder]] = (
+        Field(
+            None,
+            description='Component decoding the response so records can be extracted.',
+            title='Decoder',
+        )
     )
     parameters: Optional[Dict[str, Any]] = Field(None, alias='$parameters')